requirements.txt
tornado==3.2
voluptuous==0.8.1
pycrypto==2.6.1
mysql-python==1.2.5
xpinyin==0.5.0
xlrd==0.9.2
xlwt==0.7.5
redis==2.9.0
hiredis==0.1.2
rmb==0.0.3
pillow==2.3.0
